added Drowable interface and minot logic change in main method

+++ src/main/kotlin/mate/academy/model/RightTriangle.kt
@@ -1,6 +1,6 @@
 package mate.academy.model
 
-class RightTriangle(color: Color, val firstLeg: Double, val secondLeg: Double, ) : Figure(color) {
+class RightTriangle(color: Color, val firstLeg: Int, val secondLeg: Int, ) : Figure(color) {
     override fun draw() {
         println("Figure: right triangle, first leg: $firstLeg, second leg: $secondLeg, color: $color")
     }
